pyproject.toml
[tool.poetry]
name = "trading-platform-comment-analysis"
version = "0.1.0"
description = ""
authors = ["mehulklh"]
license = "GPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.0.1"
pandas = "^2.2.2"
matplotlib = "^3.9.1"
praw = "^7.7.1"
sqlalchemy = "^2.0.31"
googlesearch-python = "^1.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
